+++ src/main/java/com/example/Compras/dto/CiudadResponseDTO.java
@@ -1,13 +1,22 @@
 package com.example.Compras.dto;
 
 public class CiudadResponseDTO {
-    private Integer id;
+
+    private Long id;      // ✅ corregido a Long
     private String nombre;
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() {
+        return id;
+    }
+    public void setId(Long id) {
+        this.id = id;
+    }
 
-    public String getNombre() { return nombre; }
-    public void setNombre(String nombre) { this.nombre = nombre; }
-
+    public String getNombre() {
+        return nombre;
+    }
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
 }
+
