+++ src/main/java/com/example/Compras/repositories/ImpuestoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ImpuestoRepository extends JpaRepository<Impuesto, Integer> {
+public interface ImpuestoRepository extends JpaRepository<Impuesto, Long> {
     Optional<Impuesto> findByNombreIgnoreCase(String nombre);
 }
